Moves Deletable to reflex.
